+++ frontend/src/components/trip/TripItemAddModal/TripItemExpenseInput/TripItemExpenseInput.tsx
@@ -1,3 +1,4 @@
+import { useTripItemExpenseInput } from '@/hooks/useTripItemExpenseInput';
 import { CURRENCY_ICON, DEFAULT_CURRENCY } from '@constants/trip';
 import { useQueryClient } from '@tanstack/react-query';
 import type { ExpenseCategoryData } from '@type/expense';
@@ -21,48 +22,8 @@ const TripItemExpenseInput = ({
   initialExpenseValue,
   updateInputValue,
 }: TripItemExpenseInputProps) => {
-  const [_, setExpenseValue] = useState<TripItemFormType['expense']>(
-    initialExpenseValue ?? {
-      currency: DEFAULT_CURRENCY,
-      categoryId: 100,
-      amount: 0,
-    }
-  );
-
-  const queryClient = useQueryClient();
-  const expenseCategoryData = queryClient.getQueryData<ExpenseCategoryData[]>(['expenseCategory'])!;
-
-  const handleCategoryChange = (event: ChangeEvent<HTMLSelectElement>) => {
-    setExpenseValue((prevExpenseValue) => {
-      const newExpenseValue = { ...prevExpenseValue!, categoryId: Number(event.target.value) };
-
-      if (prevExpenseValue?.amount) updateInputValue('expense', newExpenseValue);
-
-      return newExpenseValue;
-    });
-  };
-
-  const handleCurrencyChange = (event: ChangeEvent<HTMLSelectElement>) => {
-    setExpenseValue((prevExpenseValue) => {
-      const newExpenseValue = { ...prevExpenseValue!, currency: event.target.value };
-
-      if (prevExpenseValue?.amount) updateInputValue('expense', newExpenseValue);
-
-      return newExpenseValue;
-    });
-  };
-
-  const handleAmountChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setExpenseValue((prevExpenseValue) => {
-      const newExpenseValue = {
-        ...prevExpenseValue!,
-        amount: Number(event.target.value),
-      };
-      updateInputValue('expense', newExpenseValue.amount === 0 ? null : newExpenseValue);
-
-      return newExpenseValue;
-    });
-  };
+  const { expenseCategoryData, handleCategoryChange, handleCurrencyChange, handleAmountChange } =
+    useTripItemExpenseInput(updateInputValue, initialExpenseValue);
 
   return (
     <Flex styles={{ direction: 'column', gap: Theme.spacer.spacing2 }}>
